+++ sisexperto/sisexperto/Migrations/201211052112424_nuevo.cs
@@ -7,7 +7,7 @@ namespace sisExperto.Migrations
     {
         public override void Up()
         {
-            AddColumn("dbo.ExpertosPonderadosEnProyectos", "Peso", c => c.Double(nullable: false));
+            AddColumn("dbo.ExpertosPonderadosEnProyectos", "Peso", c => c.Int(nullable: false));
         }
         
         public override void Down()
